+++ lib/flutter/vesper_player_performance_diagnostics/ios/vesper_player_performance_diagnostics/Package.swift
@@ -1,7 +1,7 @@
 // swift-tools-version: 5.9
 import PackageDescription
 
-private let vesperPlayerKitVersion: Version = "0.6.0"
+private let vesperPlayerKitVersion: Version = "0.6.1"
 
 let package = Package(
     name: "vesper_player_performance_diagnostics",